Consolidated the reading/writing of miscInfo, CoordSys and ImSize info of the complex grid and CFs in the ImageInformation class. ImageInformation class is now used to also read these from records.  If these record-fields aren't found in the CFC (e.g. in older CFCs) it falls back to using the ImageInterface route (the standard route, which is slower)
This is has been tested to (a) make a blank CFC and fill it, (b) against existing (old) CFC via roadrunner.

The next step in closing those work is to use ImageInformation class to save the information if it does not exist.  This way
older (existing) CFCs will also get this information after this code is applied once.