AWConvFunc::makeConvFunction2() and ::fillConvFunction2() methods now should work if "uvgrid.im" is not in the CFC. Changed the interface of ::fillConvFunction2 to accept a pointer to skyImage_l -- in prep for deprecating this parameter alltogether.
The logic encoded if ImageInformation<T> is constructed without error, the CFC must be a new one.  If not, then "uvgrid.im"
should exist (the older CFCs), which then is used to save the ImageInformation<T> and thus convert the CFC to the new format.
AFAICT, this logic should cover all existing CFCs and any new-format CFC produced.