Nominally tested code for fast CFC storing.
This required changes in the "blank" CFC format.  Now a blank CFC only
holds the meta data of the CFs as Records, and those are written to
disk in a multi-threaded code that parallelizes the loop over the CFB
Frequency axis.

The changes are in Coyote_func.h, CFCell.{cc,h} AWCF.cc CFB.{cc,h} and
CFS2.{cc,h}