Fix in cmake/Findspackenv.cmake.
Tested it by setting SPACK_ROOT but not SPACK_ENV. Warnning message
was issued and SPACKENV_FOUND was set to FALSE.

With both SPACK_ROOT and SPACK_ENV set, STATUS message was issued with
SPACK_FLAGS set correctly and SPACKENV_FOUND set to TRUE.

Also checked that the SPACK_FLAGS *cmake variable* was available in
the cmake/*/*.cmake files that use it.

SPACK_ROOT and SPACK_ENV are *shell environment variables* and can be
set by-hand, or by activating an spack env.  SPACK_ROOT remains set
even after deactivating spack env, but not SPACK_ENV.  Hence the logic
in the current code.

The code in the version before this had unset variables, and variables
that aren't used or useful here.  SPACK_FLAGS wasn't correctly set
either.